requirements.txt
openai
requests
python-dotenv
tkinter
